Add new services

+++ chats/api/src/main/java/com/chats/api/http/rest/ChatRestController.java
@@ -9,6 +9,7 @@ import com.chats.api.dto.ChatCreateEditDto;
 import com.chats.api.dto.ChatReadDto;
 
 import com.chats.api.service.ChatService;
+import com.chats.api.http.connection.UsersUrlConnection;
 
 import lombok.RequiredArgsConstructor;
 
@@ -27,6 +28,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 
 
 @RestController
@@ -38,35 +40,62 @@ public class ChatRestController {
     private final ChatService chatService;
     
     @GetMapping("/")
-    public List<ChatReadDto> findAll() {
+    public List<ChatReadDto> findAll(@RequestHeader("Sessionid") String sessionid) {
         return chatService.findAll();
     }
     
     @GetMapping("/{id}/")
-    public ChatReadDto findById(@PathVariable Long id) {
+    public ChatReadDto findById(@PathVariable Long id, @RequestHeader("Sessionid") String sessionid) {
         return chatService.findById(id)
             .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
     }
     
     @PostMapping("/")
     @ResponseStatus(HttpStatus.CREATED)
-    public ChatReadDto create(@Validated @RequestBody ChatCreateEditDto chatDto) {
-        return chatService.create(chatDto);
+    public ResponseEntity<?> create(@Validated @RequestBody ChatCreateEditDto chatDto,
+                            @RequestHeader("Sessionid") String sessionid) {
+        try { 
+            boolean is_current = UsersUrlConnection.is_current(sessionid, chatDto.getUser1());
+            if (is_current) {
+                return new ResponseEntity<>(chatService.create(chatDto), HttpStatus.CREATED);
+            }
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        } 
+        catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+       
     }
     
     @PutMapping("/{id}/")
-    public ChatReadDto update(@PathVariable Long id, @Validated @RequestBody ChatCreateEditDto chatDto) {
-        return chatService.update(id, chatDto)
-            .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+    public ResponseEntity<?> update(@PathVariable Long id, @Validated @RequestBody ChatCreateEditDto chatDto,
+                                @RequestHeader("Sessionid") String sessionid) {
+        try {
+            boolean is_current = UsersUrlConnection.is_current(sessionid, chatDto.getUser1());
+            if (is_current) {
+                return new ResponseEntity<>(chatService.update(id, chatDto)
+                    .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND)),
+                    HttpStatus.OK);
+            }
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
     }
 
     @DeleteMapping("/{id}/")
-    public ResponseEntity<?> delete(@PathVariable Long id) {
-        return chatService.delete(id)
-        ?
-        noContent().build()
-        :
-        notFound().build();
+    public ResponseEntity<?> delete(@PathVariable Long id, @RequestHeader("Sessionid") String sessionid) {
+        try {
+            boolean is_current = UsersUrlConnection.is_current(sessionid, chatService.findById(id).get().getUser1());
+            if (is_current) {
+                return chatService.delete(id)
+                    ? noContent().build()
+                    : notFound().build();
+            }
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
     }
 
 }
