+++ tasks/api/src/main/java/com/tasks/api/http/rest/TaskRestContoller.java
@@ -7,6 +7,7 @@ import com.tasks.api.dto.PageResponse;
 import com.tasks.api.dto.TaskCreateEditDto;
 import com.tasks.api.dto.TaskFilter;
 import com.tasks.api.dto.TaskReadDto;
+import com.tasks.api.http.connection.UsersUrlConnection;
 import com.tasks.api.service.TaskService;
 
 import lombok.RequiredArgsConstructor;
@@ -28,6 +29,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 
 @RestController
 @CrossOrigin(origins = {"http://localhost:3000", "http://127.0.0.1:3000"})
@@ -51,20 +53,51 @@ public class TaskRestContoller {
     
     @PostMapping("/")
     @ResponseStatus(HttpStatus.CREATED)
-    public TaskReadDto create(@Validated @RequestBody TaskCreateEditDto task) {
-        return taskService.create(task);
+    public ResponseEntity<?> create(@Validated @RequestBody TaskCreateEditDto task,
+                                @RequestHeader("Sessionid") String sessionid) {
+        try { 
+            boolean is_current = UsersUrlConnection.is_current(sessionid, task.getOf());
+            if (is_current) {
+                return new ResponseEntity<>(taskService.create(task), HttpStatus.CREATED);
+            }
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        } 
+        catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        
     }
     
     @PutMapping("/{id}")
-    public TaskReadDto update(@PathVariable Integer id, @Validated @RequestBody TaskCreateEditDto task) {
-        return taskService.update(id, task)
-            .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+    public ResponseEntity<?> update(@PathVariable Integer id, @Validated @RequestBody TaskCreateEditDto task,
+                        @RequestHeader("Sessionid") String sessionid) {
+        try {
+            boolean is_current = UsersUrlConnection.is_current(sessionid, task.getOf());
+            if (is_current) {
+                return new ResponseEntity<>(taskService.update(id, task)
+                    .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND)),
+                    HttpStatus.OK);
+            }
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
     }
     
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> delete(@PathVariable Integer id) {
-        return taskService.delete(id)
-            ? noContent().build()
-            : notFound().build();
+    public ResponseEntity<?> delete(@PathVariable Integer id, 
+                                @RequestHeader("Sessionid") String sessionid) {
+        try {
+            boolean is_current = UsersUrlConnection.is_current(sessionid, taskService.findById(id).get().getOf());
+            if (is_current) {
+                return taskService.delete(id)
+                    ? noContent().build()
+                    : notFound().build();
+            }
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        
     }
 }
